+++ bitalarm-impl/src/test/java/com/dwarfeng/bitalarm/impl/service/AlarmTypeIndicatorMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -31,8 +32,7 @@ public class AlarmTypeIndicatorMaintainServiceImplTest {
         alarmTypeIndicators = new ArrayList<>();
         for (int i = 0; i < 5; i++) {
             AlarmTypeIndicator alarmTypeIndicator = new AlarmTypeIndicator(
-                    new StringIdKey("test.alarm_type_indicator." + i),
-                    "测试 " + i, "remark"
+                    new StringIdKey("test.alarm_type_indicator." + i), "测试 " + i, "remark"
             );
             alarmTypeIndicators.add(alarmTypeIndicator);
         }
@@ -54,6 +54,9 @@ public class AlarmTypeIndicatorMaintainServiceImplTest {
             }
         } finally {
             for (AlarmTypeIndicator alarmTypeIndicator : alarmTypeIndicators) {
+                if (Objects.isNull(alarmTypeIndicator)) {
+                    continue;
+                }
                 alarmTypeIndicatorMaintainService.deleteIfExists(alarmTypeIndicator.getKey());
             }
         }
