+++ bitalarm-impl/src/test/java/com/dwarfeng/bitalarm/impl/service/CurrentAlarmMaintainServiceImplTest.java
@@ -17,6 +17,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.Date;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -37,28 +38,11 @@ public class CurrentAlarmMaintainServiceImplTest {
 
     @Before
     public void setUp() {
-        parentPoint = new Point(
-                new LongIdKey(1),
-                "test-point",
-                "test-point"
-        );
+        parentPoint = new Point(new LongIdKey(1), "test-point", "test-point");
         parentAlarmSetting = new AlarmSetting(
-                null,
-                parentPoint.getKey(),
-                true,
-                1,
-                "我是报警信息",
-                "alarmType",
-                "测试用报警设置"
+                null, parentPoint.getKey(), true, 1, "我是报警信息", "alarmType", "测试用报警设置"
         );
-        currentAlarm = new CurrentAlarm(
-                null,
-                parentPoint.getKey(),
-                1,
-                "我是报警信息",
-                "alarmType",
-                new Date()
-        );
+        currentAlarm = new CurrentAlarm(null, parentPoint.getKey(), 1, "我是报警信息", "alarmType", new Date());
     }
 
     @After
@@ -78,9 +62,15 @@ public class CurrentAlarmMaintainServiceImplTest {
             CurrentAlarm currentAlarm1 = currentAlarmMaintainService.get(this.currentAlarm.getKey());
             assertEquals(BeanUtils.describe(this.currentAlarm), BeanUtils.describe(currentAlarm1));
         } finally {
-            pointMaintainService.deleteIfExists(parentPoint.getKey());
-            currentAlarmMaintainService.deleteIfExists(currentAlarm.getKey());
-            alarmSettingMaintainService.deleteIfExists(parentAlarmSetting.getKey());
+            if (Objects.nonNull(currentAlarm.getKey())) {
+                currentAlarmMaintainService.deleteIfExists(currentAlarm.getKey());
+            }
+            if (Objects.nonNull(parentAlarmSetting.getKey())) {
+                alarmSettingMaintainService.deleteIfExists(parentAlarmSetting.getKey());
+            }
+            if (Objects.nonNull(parentPoint.getKey())) {
+                pointMaintainService.deleteIfExists(parentPoint.getKey());
+            }
         }
     }
 }
